Use the same output rules for both meshed and not meshed networks
