+++ mirrord/agent/src/steal/ip_tables/output.rs
@@ -1,0 +1,87 @@
+use std::sync::Arc;
+
+use async_trait::async_trait;
+use mirrord_protocol::Port;
+use nix::unistd::getgid;
+use tracing::warn;
+
+use crate::{
+    error::Result,
+    steal::ip_tables::{chain::IPTableChain, IPTables, Redirect},
+};
+
+pub(crate) struct OutputRedirect<IPT: IPTables> {
+    pub(crate) managed: IPTableChain<IPT>,
+}
+
+impl<IPT> OutputRedirect<IPT>
+where
+    IPT: IPTables,
+{
+    const ENTRYPOINT: &'static str = "OUTPUT";
+
+    pub fn create(ipt: Arc<IPT>, chain_name: String) -> Result<Self> {
+        let managed = IPTableChain::create(ipt, chain_name)?;
+
+        let gid = getgid();
+        managed
+            .add_rule(&format!("-m owner --gid-owner {gid} -p tcp -j RETURN"))
+            .inspect_err(|_| {
+                warn!("Unable to create iptable rule with \"--gid-owner {gid}\" filter")
+            })?;
+
+        Ok(OutputRedirect { managed })
+    }
+
+    pub fn load(ipt: Arc<IPT>, chain_name: String) -> Result<Self> {
+        let managed = IPTableChain::create(ipt, chain_name)?;
+
+        Ok(OutputRedirect { managed })
+    }
+}
+
+/// This wrapper adds a new rule to the NAT OUTPUT chain to redirect "localhost" traffic as well
+/// Note: OUTPUT chain is only traversed for packets produced by local applications
+#[async_trait]
+impl<IPT> Redirect for OutputRedirect<IPT>
+where
+    IPT: IPTables + Send + Sync,
+{
+    async fn mount_entrypoint(&self) -> Result<()> {
+        self.managed.inner().add_rule(
+            Self::ENTRYPOINT,
+            &format!("-j {}", self.managed.chain_name()),
+        )?;
+
+        Ok(())
+    }
+
+    async fn unmount_entrypoint(&self) -> Result<()> {
+        self.managed.inner().remove_rule(
+            Self::ENTRYPOINT,
+            &format!("-j {}", self.managed.chain_name()),
+        )?;
+
+        Ok(())
+    }
+
+    async fn add_redirect(&self, redirected_port: Port, target_port: Port) -> Result<()> {
+        let redirect_rule = format!(
+            "-o lo -m tcp -p tcp --dport {redirected_port} -j REDIRECT --to-ports {target_port}"
+        );
+
+        self.managed.add_rule(&redirect_rule)?;
+
+        Ok(())
+    }
+
+    async fn remove_redirect(&self, redirected_port: Port, target_port: Port) -> Result<()> {
+        let redirect_rule = format!(
+            "-o lo -m tcp -p tcp --dport {redirected_port} -j REDIRECT --to-ports {target_port}"
+        );
+
+        self.managed.remove_rule(&redirect_rule)?;
+
+        Ok(())
+    }
+}
